requirements.txt
# Core Dependencies for Legal Events v2
# Python 3.12+

# FastAPI and Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
sse-starlette>=1.6.5

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Storage (MinIO)
minio>=7.2.0

# Queue (Redis/RQ)
redis>=5.0.0
rq>=1.15.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Data Processing  
pandas>=2.2.0
openpyxl>=3.1.0
numpy<2

# Existing Pipeline Dependencies
# (These are needed to run the extraction pipeline)
docling>=0.1.0
anthropic>=0.40.0
openai>=1.0.0
google-generativeai>=0.8.0
requests>=2.32.0
python-dateutil>=2.9.0

# CORS and Security
python-multipart>=0.0.6

# Development/Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Monitoring (optional)
prometheus-client>=0.19.0
